Fix: Wrap useSearchParams in Suspense boundary for static generation - fixes Vercel build failure
